moved newly created classes to Classes file; created new file to avoid cyclic imports
