+++ code/drasil.cabal
@@ -42,6 +42,7 @@ library
     , Language.Drasil.Expr.Precedence
     , Language.Drasil.Classes
     , Language.Drasil.Label
+    , Language.Drasil.LabelHelpers
     , Language.Drasil.Config
     , Language.Drasil.Document
     , Language.Drasil.Format
